+++ Sourcery/main.py
@@ -67,8 +67,9 @@ def main():
         default=True, help="Make reliability density plot. Default is True."
         " To disable add -dmp on the command line.")
  
-    add("-rel", "--rel-thresh", dest="rel_thresh", default=None, type=float, 
-        help= "Sets a reliability threshold. Default is None.")
+    add("-drel", "--reset-rel", dest="reset_rel", action="store_true",
+        default=False, help=" Assigns R=0 for sources with cf < 0.002 and" 
+        "R > 0.60.")
 
     add("-beam", "--beam-cluster", dest="do_beam", default=False,
         action="store_true", help= "Increases the Gaussian groupings by 20 percent the"
@@ -82,7 +83,7 @@ def main():
         help="The area to compute the local variance, in beam sizes."
         " Default value is 10 beam sizes.")
      
-    add("-rel_rm", "--rel-sources-excl", dest="rel_src_excl", 
+    add("-rel_rm", "--rel-rmsrc", dest="rel_rmsrc", 
         action="append", default=None, help="Delete sources within a radius;"
         " e.g ra, dec, radius (in degrees). For more than"
         " one region: ra1,dec1,radius1:ra2,dec2,radius2. Default is None.")
@@ -131,7 +132,7 @@ def main():
 
     add("-phrm", "--phasecenter-remove", dest="phase_center_rm",
         type=float, default=None, help="The radius excluded from"
-        " direction-dependent source selection. NB: this radius is wrt to"
+        " direction-dependent source selection. NB: this radius is w.r.t to"
         " the phase center. Default is None.")
 
     add('-jc', '--json-config', dest='config', default=None,
@@ -148,13 +149,9 @@ def main():
     pybdsm_opts = dict([ items.split("=") for items in args.to_pybdsm ] ) \
                        if args.to_pybdsm else {}
 
-    if args.rel_src_excl:
-        rel_rmsrc = args.rel_src_excl[0].split(':')
-    else:
-        rel_rmsrc = None
-
     if not args.image:
         print("ATTENTION: No image provided. Aborting")
+        
 
     # making outdirectory
     def get_prefix(prefix, imagename, outdir):
@@ -268,11 +265,11 @@ def main():
                      args.source_finder, makeplots=args.do_relplots, 
                      do_psf_corr=args.add_psfcorr, do_local_var=args.add_locvar,
                      psf_corr_region=psfregion, local_var_region=locregion, 
-                     rel_excl_src=rel_rmsrc, pos_smooth=args.pos_smooth,
+                     rel_excl_src=args.rel_rmsrc, pos_smooth=args.pos_smooth,
                      neg_smooth=args.neg_smooth, loglevel=args.log_level, 
                      thresh_isl=args.thresh_isl, thresh_pix=args.thresh_pix,
                      neg_thresh_isl=args.neg_thresh_isl, neg_thresh_pix=
-                     args.neg_thresh_pix, prefix=prefix,  
+                     args.neg_thresh_pix, prefix=prefix, reset_rel=args.reset_rel, 
                      do_nearsources=args.do_nearsources, increase_beam_cluster=
                      args.do_beam, savemask_neg=args.savemask_neg,
                      savemask_pos=args.savemask_pos, **pybdsm_opts)