+++ Sourcery/CHANGES.md
@@ -1,6 +1,9 @@
-##Current version: 0.2.8
+##Current version: 0.2.9
 
 
-##From version 0.2.7 to 0.2.8, the changes made were;
+##From version 0.2.8 to 0.2.9, the changes made were;
 
-#### Threshold for local variance is 0.8 not 0.4. The local and PSF correlation are computed inside reliability script to solve assigning tags to wrong sources. 
+#### Shapes are corrected: sourcery able to assign pnt and Gau appropriately, and uses DC_Maj/Min as a model and Maj/Min for reliability computation.
+#### The flux estimations are corrected. Uses masked image to create islands and the actual image for detection (model fitting)
+#### Takes the radius of sources to exclude inside the reliability script (previously it was done inside main-- had problems)
+#### Has an option to remove sources with correlation < 0.002 and reliability > 0.60